bug_:fix

+++ offers/models.py
@@ -35,7 +35,7 @@ class OfferDetails(models.Model):
     title = models.CharField(max_length=255)
     revisions = models.IntegerField()
     delivery_time_in_days = models.IntegerField()
-    price = models.IntegerField()
+    price = models.FloatField()
     features = models.JSONField(default=list)
     offer_type = models.CharField(max_length=8, choices=PAKET_TYPES)
